Code review changes (2)
