+++ lib/src/ast/sass/expression/interpolated_function.dart
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc. Use of this source code is governed by an
+// Copyright 2021 Google Inc. Use of this source code is governed by an
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
