+++ lib/src/utils.dart
@@ -416,73 +416,6 @@ int consumeEscapedCharacter(StringScanner scanner) {
   }
 }
 
-extension SpanExtensions on FileSpan {
-  /// Returns this span with all whitespace trimmed from both sides.
-  FileSpan trim() => trimLeft().trimRight();
-
-  /// Returns this span with all leading whitespace trimmed.
-  FileSpan trimLeft() {
-    var start = 0;
-    while (isWhitespace(text.codeUnitAt(start))) {
-      start++;
-    }
-    return subspan(start);
-  }
-
-  /// Returns this span with all trailing whitespace trimmed.
-  FileSpan trimRight() {
-    var end = text.length - 1;
-    while (isWhitespace(text.codeUnitAt(end))) {
-      end--;
-    }
-    return subspan(0, end + 1);
-  }
-
-  /// Returns the span of the identifier at the start of this span.
-  ///
-  /// If [includeLeading] is greater than 0, that many additional characters
-  /// will be included from the start of this span before looking for an
-  /// identifier.
-  FileSpan initialIdentifier({int includeLeading = 0}) {
-    var scanner = StringScanner(text);
-    for (var i = 0; i < includeLeading; i++) {
-      scanner.readChar();
-    }
-    _scanIdentifier(scanner);
-    return subspan(0, scanner.position);
-  }
-
-  /// Returns a subspan excluding the identifier at the start of this span.
-  FileSpan withoutInitialIdentifier() {
-    var scanner = StringScanner(text);
-    _scanIdentifier(scanner);
-    return subspan(scanner.position);
-  }
-
-  /// Returns a subspan excluding an initial at-rule and any whitespace after
-  /// it.
-  FileSpan withoutInitialAtRule() {
-    var scanner = StringScanner(text);
-    scanner.expectChar($at);
-    _scanIdentifier(scanner);
-    return subspan(scanner.position).trimLeft();
-  }
-}
-
-/// Consumes an identifier from [scanner].
-void _scanIdentifier(StringScanner scanner) {
-  while (!scanner.isDone) {
-    var char = scanner.peekChar()!;
-    if (char == $backslash) {
-      consumeEscapedCharacter(scanner);
-    } else if (isName(char)) {
-      scanner.readChar();
-    } else {
-      break;
-    }
-  }
-}
-
 extension MapExtension<K, V> on Map<K, V> {
   /// If [this] doesn't contain the given [key], sets that key to [value] and
   /// returns it.
